+++ di/src/main/java/woowacourse/di/DIActivity.kt
@@ -2,14 +2,9 @@ package woowacourse.di
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import woowacourse.di.annotation.InjectField
-import kotlin.reflect.KMutableProperty
-import kotlin.reflect.full.declaredMemberProperties
-import kotlin.reflect.full.hasAnnotation
-import kotlin.reflect.jvm.isAccessible
-import kotlin.reflect.jvm.jvmErasure
+import woowacourse.di.module.AndroidModule
 
-abstract class DIActivity(private val module: Module) : AppCompatActivity() {
+abstract class DIActivity(private val module: AndroidModule) : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -17,15 +12,6 @@ abstract class DIActivity(private val module: Module) : AppCompatActivity() {
         module.setModuleContext(this)
         val container = Container(module)
 
-        val properties =
-            this::class.declaredMemberProperties.filter { it.hasAnnotation<InjectField>() }
-        properties.forEach { property ->
-            property.isAccessible = true
-            val injectValue = container.getInstance(property.returnType.jvmErasure)
-            (property as KMutableProperty<*>).setter.call(this, injectValue)
-        }
-
-//        val injector = Injector(container)
-//        injector.injectField(this::class)
+        injectProperties(container)
     }
 }
