+++ di/src/main/java/woowacourse/di/annotation/Singleton.kt
@@ -1,3 +1,4 @@
 package woowacourse.di.annotation
 
+@Target(AnnotationTarget.FUNCTION)
 annotation class Singleton
