+++ di/src/main/java/woowacourse/di/module/AndroidModule.kt
@@ -1,8 +1,8 @@
-package woowacourse.di
+package woowacourse.di.module
 
 import android.content.Context
 
-interface Module {
+interface AndroidModule : Module {
     var context: Context
     fun setModuleContext(context: Context)
 }
