+++ di/src/main/java/woowacourse/di/module/Module.kt
@@ -1,0 +1,3 @@
+package woowacourse.di.module
+
+interface Module
