+++ di/src/test/java/woowacourse/di/FakeModule.kt
@@ -1,16 +1,10 @@
 package woowacourse.di
 
-import android.content.Context
 import woowacourse.di.annotation.InjectField
 import woowacourse.di.annotation.Singleton
+import woowacourse.di.module.Module
 
 class FakeModule : Module {
-
-    override lateinit var context: Context
-
-    override fun setModuleContext(context: Context) {
-        this.context = context
-    }
 
     fun provideFakeProductRepository(): FakeProductRepository = DefaultFakeProductRepository()
